requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
colorama==0.4.4
Django==3.2.6
django-cors-headers==3.7.0
django-extra-settings==0.1.3
django-modeltranslation==0.17.3
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
python-decouple==3.5
pytz==2021.1
regex==2021.8.3
six==1.16.0
sqlparse==0.4.1
tomli==1.2.1
